requirements.txt
# Core dependencies for bumblebee_bplusplus project

# PyTorch and vision libraries
torch>=2.0.0
torchvision>=0.15.0

# Data processing and utilities
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.0.0
opencv-python>=4.6.0

# Machine learning and metrics
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Visualization and reporting
matplotlib>=3.5.0
tabulate>=0.9.0
ftfy>=6.1.1
regex>=2023.0

# Progress tracking
tqdm>=4.62.0

# API and web requests
requests
# OpenAI API - for synthetic image generation with DALL-E 3 and GPT-4 Vision
openai>=1.0.0

# Ultralytics (for YOLO/object detection models)
ultralytics>=8.0.0

# Utilities
pyyaml>=6.0
validators>=0.20.0
python-dotenv>=1.0.0  # For loading environment variables from .env file

# Biodiversity Plus Plus - ML pipeline for species classification
bplusplus>=1.2.5
